Move mainLoop without C-z to windows-specific file
