+++ boltbrowser.go
@@ -17,27 +17,6 @@ var databaseFile string
 var db *bolt.DB
 var memBolt *BoltDB
 
-func mainLoop(memBolt *BoltDB, style Style) {
-	screens := defaultScreensForData(memBolt)
-	displayScreen := screens[BrowserScreenIndex]
-	layoutAndDrawScreen(displayScreen, style)
-	for {
-		event := termbox.PollEvent()
-		if event.Type == termbox.EventKey {
-			newScreenIndex := displayScreen.handleKeyEvent(event)
-			if newScreenIndex < len(screens) {
-				displayScreen = screens[newScreenIndex]
-				layoutAndDrawScreen(displayScreen, style)
-			} else {
-				break
-			}
-		}
-		if event.Type == termbox.EventResize {
-			layoutAndDrawScreen(displayScreen, style)
-		}
-	}
-}
-
 func main() {
 	var err error
